Move `WeakRefVirtualProxy` to its own file
